requirements.txt
typer[all]
rich
gitpython
openai
anthropic
google-generativeai
ollama
python-dotenv
pydantic
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
jinja2
pypdf
